Fix to search bq and hl parameters
The existing code takes the example config options (as per https://github.com/AtlasOfLivingAustralia/ala-install/blob/master/ansible/roles/bie-index/templates/bie-index-config.properties) and splits them up into individual letters, which doesn't work. Is it possible to either update the ansible template to reflect the way in which these parameters must now be provided to work as arrays/lists as implied in the original code, or basically revert to the code in this fork? Thanks guys